Uses abstract digest logic in sealed class.
Derives all parameters from digest with algorithm tag supplied per
useful algorithm.

Adds support for SHA2, SHA3, Keccak and Blake2 family of functions.

+++ common/src/main/kotlin/pt/um/masb/common/hash/Hasher.kt
@@ -7,7 +7,7 @@ package pt.um.masb.common.hash
  */
 interface Hasher {
     val id: Hash
-    val hashSize: Long
+    val hashSize: Int
 
 
     /**
